# apps/api/src/openapi.yaml
openapi: 3.0.3
info:
  title: QiFlow API
  description: |
    The QiFlow payment gateway API. Integrate once, accept Qi everywhere.

    ## Authentication

    QiFlow uses two authentication mechanisms depending on the endpoint group:

    ### JWT Bearer Token (Dashboard endpoints)
    Issued on login / register. Short-lived access token (15 min) + long-lived refresh token (7 days).
    ```
    Authorization: Bearer <accessToken>
    ```

    ### API Key (Programmatic / v1 endpoints)
    Generated from the dashboard or via `POST /merchants/me/api-keys`.
    ```
    X-API-Key: qf_test_xxxxxxxxxxxxxxxx
    ```

    ## Rate Limiting
    Rate-limit headers are returned on every response:
    | Header | Description |
    |---|---|
    | `RateLimit-Limit` | Max requests per window |
    | `RateLimit-Remaining` | Requests left in current window |
    | `RateLimit-Reset` | Seconds until window resets |

    Auth endpoints have stricter limits (5 req/min for login, 3 req/min for register).
    All other endpoints share a global limit of 500 req/15 min.

    ## Error Format
    All errors follow a consistent shape:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable message"
      }
    }
    ```
  version: 1.0.0
  contact:
    name: QiFlow
    url: https://qiflow.xyz
servers:
  - url: https://api.qiflow.xyz
    description: Production
  - url: http://localhost:3001
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from `/auth/login` or `/auth/register`
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key obtained from the merchant dashboard or `POST /merchants/me/api-keys`

  # ── Reusable Schemas ─────────────────────────────────────────────────────────
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: INVALID_CREDENTIALS
            message:
              type: string
              example: Invalid email or password
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        details:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email: ["Invalid email address"]
            password: ["Password must be at least 8 characters"]

    # ── Tokens ───────────────────────────────────────────────────────────────────
...
    Payment:
      type: object
      required: [id, paymentCode, amount, currency, status, receivingAddress, expiresAt, createdAt]
      properties:
        id:
          type: string
          format: uuid
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        paymentCode:
          type: string
          example: pay_82hd91
          description: Unique short code used in checkout URLs
        amount:
          type: string
          description: Payment amount in Qi (decimal string)
          example: "50.00000000"
        currency:
          type: string
          example: QI
        description:
          type: string
          nullable: true
          example: Nike Air Force 1
        status:
          type: string
          enum: [CREATED, PENDING, PROCESSING, COMPLETED, FAILED, CANCELLED, EXPIRED]
          example: PENDING
        receivingAddress:
          type: string
          description: Quai wallet address that receives the payment
          example: "0x00473a216f2b1d382759e612bf6029fa037e95b2"
        txHash:
          type: string
          nullable: true
          description: Quai transaction hash once payment is broadcast
          example: "0xabc123..."
        checkoutUrl:
          type: string
          description: Shareable hosted checkout URL
          example: https://qiflow.xyz/pay/pay_82hd91
        metadata:
          type: object
          nullable: true
          additionalProperties: true
          description: Arbitrary key/value pairs attached to the payment
          example:
            orderId: "1024"
            customer: "alice@example.com"
        expiresAt:
          type: string
          format: date-time
          example: "2026-01-01T01:00:00.000Z"
        completedAt:
          type: string
          format: date-time
          nullable: true
          example: null
        createdAt:
          type: string
          format: date-time
          example: "2026-01-01T00:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-01-01T00:00:00.000Z"

    # ── Webhook ──────────────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Missing or invalid Authorization header
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        RateLimit-Limit:
          schema:
            type: integer
        RateLimit-Remaining:
          schema:
            type: integer
        RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            error:
              code: RATE_LIMIT_EXCEEDED
              message: Too many requests, please try again later.

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /v1/payments:
    post:
      tags: [Payments]
      summary: Create a payment
      description: |
        Creates a new payment session and returns a hosted checkout URL.
        The merchant's `walletAddress` must be set before creating payments.

        **Authentication:** API Key (`X-API-Key` header)
      operationId: createPayment
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: string
                  description: Payment amount in Qi
                  example: "100"
                currency:
                  type: string
                  default: QI
                  example: QI
                description:
                  type: string
                  maxLength: 500
                  example: Order #1024
                metadata:
                  type: object
                  additionalProperties: true
                  description: Arbitrary key/value pairs stored with the payment
                  example:
                    orderId: "1024"
                    customer: "alice@example.com"
            example:
              amount: "100"
              currency: QI
              description: Order #1024
              metadata:
                orderId: "1024"
      responses:
        '201':
          description: Payment created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'

    get:
      tags: [Payments]
      summary: List payments
      description: |
        Returns a paginated list of payments for the authenticated merchant.

        **Authentication:** API Key (`X-API-Key` header)
      operationId: listPayments
      security:
        - ApiKeyAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [CREATED, PENDING, PROCESSING, COMPLETED, FAILED, CANCELLED, EXPIRED]
          description: Filter by payment status
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Number of results per page
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Pagination offset
      responses:
        '200':
          description: List of payments
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          payments:
                            type: array
                            items:
                              $ref: '#/components/schemas/Payment'
                          total:
                            type: integer
                            example: 42
                          limit:
                            type: integer
                            example: 20
                          offset:
                            type: integer
                            example: 0
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/payments/{id}:
    get:
      tags: [Payments]
      summary: Get a payment
      description: |
        Returns a single payment by ID or payment code.

        **Authentication:** API Key (`X-API-Key` header)
      operationId: getPayment
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Payment UUID or payment code (e.g. `pay_82hd91`)
          example: pay_82hd91
      responses:
        '200':
          description: Payment found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Payment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Webhooks ──────────────────────────────────────────────────────────────────
...